Remove/Update tornado/twister tests.
